+++ Realm/RLMSyncManager.mm
@@ -38,6 +38,7 @@ using namespace realm;
 
 // NEXT-MAJOR: All the code associated to the logger from sync manager should be removed.
 using Level = realm::util::Logger::Level;
+using LogCategory = realm::util::LogCategory;
 
 namespace {
 Level levelForSyncLogLevel(RLMSyncLogLevel logLevel) {
@@ -73,14 +74,14 @@ RLMSyncLogLevel logLevelForLevel(Level logLevel) {
 #pragma mark - Loggers
 
 struct CocoaSyncLogger : public realm::util::Logger {
-    void do_log(const realm::util::LogCategory&, Level, const std::string& message) override {
-        NSLog(@"Sync: %@", RLMStringDataToNSString(message));
+    void do_log(const realm::util::LogCategory& category, Level, const std::string& message) override {
+        NSLog(@"%@: %@", RLMStringDataToNSString(category.get_name()), RLMStringDataToNSString(message));
     }
 };
 
 static std::unique_ptr<realm::util::Logger> defaultSyncLogger(realm::util::Logger::Level level) {
     auto logger = std::make_unique<CocoaSyncLogger>();
-    logger->set_level_threshold(level);
+    logger->set_level_threshold(LogCategory::sync, level);
     return std::move(logger);
 }
 
@@ -154,7 +155,7 @@ std::shared_ptr<realm::util::Logger> RLMWrapLogFunction(RLMSyncLogFunction fn) {
         _syncManager->set_logger_factory([logFn](realm::util::Logger::Level level) {
             auto logger = std::make_unique<CallbackLogger>();
             logger->logFn = logFn;
-            logger->set_level_threshold(level);
+            logger->set_level_threshold(LogCategory::sync, level);
             return logger;
         });
     }